- Segmented everything into different libraries for readability and future precompile plans - Got rid of some cache files getting past the gitignore (since when where those files there wtf?????) Protocol I2C: - Fixed readOneMessage not reading CRC if package len was 0 - Fixed writeData not accounting for the bytes it has already sent when reading the buffer - sendQueue deadline fixed Hardware Node: - Added radsToSteps and StepstoRads functions to avoid long inline operation - Fixed possible race condition with the input ros callbacks
