+++ workspace/src/hardware/CMakeLists.txt
@@ -5,7 +5,6 @@ if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
-
 find_package(ament_cmake        REQUIRED)
 find_package(ament_cmake_python REQUIRED)
 find_package(rclpy              REQUIRED)
@@ -17,7 +16,9 @@ find_package(trajectory_msgs    REQUIRED)
 find_package(sensor_msgs        REQUIRED)
 find_package(rosidl_default_generators REQUIRED)
 
-
+# ---------------------------------------------------------------------------
+# Custom message generation
+# ---------------------------------------------------------------------------
 rosidl_generate_interfaces(${PROJECT_NAME}
   "msg/JointControl.msg"
   DEPENDENCIES std_msgs
@@ -25,7 +26,15 @@ rosidl_generate_interfaces(${PROJECT_NAME}
 
 ament_export_dependencies(rosidl_default_runtime)
 
+# ---------------------------------------------------------------------------
+# dc_motors executable
+# ---------------------------------------------------------------------------
 add_executable(dc_motors src/dc_motors.cpp)
+
+# Expose the include/ directory so headers are found without path prefixes
+target_include_directories(dc_motors PRIVATE
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+)
 
 ament_target_dependencies(dc_motors
   rclcpp
@@ -40,15 +49,12 @@ rosidl_target_interfaces(dc_motors
   ${PROJECT_NAME} "rosidl_typesupport_cpp"
 )
 
-
+# ---------------------------------------------------------------------------
+# Install
+# ---------------------------------------------------------------------------
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
-  # the following line skips the linter which checks for copyrights
-  # comment the line when a copyright and license is added to all source files
   #set(ament_cmake_copyright_FOUND TRUE)
-  # the following line skips cpplint (only works in a git repo)
-  # comment the line when this package is in a git repo and when
-  # a copyright and license is added to all source files
   #set(ament_cmake_cpplint_FOUND TRUE)
   #ament_lint_auto_find_test_dependencies()
 endif()
@@ -62,6 +68,11 @@ install(PROGRAMS
   scripts/motor_node.py
   scripts/joint_mux.py
   DESTINATION lib/${PROJECT_NAME}
+)
+
+# Install headers so other packages can consume them if needed
+install(DIRECTORY include/
+  DESTINATION include/${PROJECT_NAME}
 )
 
 ament_package()
